Time: 1 ms (99.52%) | Memory: 57.3 MB (20.69%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.java
@@ -11,6 +11,6 @@ class Solution {
                 sum=0;
             }
         }
-        return (int)max;
+        return max;
     }
 }
